fix issue creating new category with correct permission

+++ category/views.py
@@ -3,6 +3,8 @@ from drf_yasg.utils import swagger_auto_schema
 from rest_framework import status, permissions
 from rest_framework.response import Response
 from rest_framework.views import APIView
+
+from wissen_api.permissions import HasInstructorProfile
 from .models import Category
 from .serializers import CategorySerializer
 
@@ -22,10 +24,18 @@ class CategoryList(APIView):
 
     @swagger_auto_schema(request_body=CategorySerializer)
     def post(self, request):
+        # Allow only instructors to create a course
+        if not HasInstructorProfile().has_permission(request, self):
+            return Response(
+                {"detail": "You must have an instructor profile to create a course."},
+                status=status.HTTP_403_FORBIDDEN
+            )
+
         serializer = CategorySerializer(
             data=request.data, context={'request': request}
         )
         if serializer.is_valid():
+            serializer.save()
             return Response(
                 serializer.data, status=status.HTTP_201_CREATED
             )
@@ -54,9 +64,15 @@ class CategoryDetails(APIView):
         serializer = CategorySerializer(category, context={'request': request})
         return Response(serializer.data)
 
-
     @swagger_auto_schema(request_body=CategorySerializer)
     def put(self, request, pk):
+        # Allow only instructors to create a course
+        if not HasInstructorProfile().has_permission(request, self):
+            return Response(
+                {"detail": "You must have an instructor profile to create a course."},
+                status=status.HTTP_403_FORBIDDEN
+            )
+
         category = self.get_object(pk)
         serializer = CategorySerializer(category, data=request.data, context={'request': request})
         if serializer.is_valid():
@@ -65,6 +81,13 @@ class CategoryDetails(APIView):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     def delete(self, request, pk):
+        # Allow only instructors to create a course
+        if not HasInstructorProfile().has_permission(request, self):
+            return Response(
+                {"detail": "You must have an instructor profile to create a course."},
+                status=status.HTTP_403_FORBIDDEN
+            )
+
         category = self.get_object(pk)
         category.delete()
         return Response(
